+++ src/content/menu/MenuBtnView.js
@@ -3,27 +3,28 @@ var Transform       = require('famous/core/Transform');
 var Modifier        = require('famous/core/Modifier');
 var Surface         = require('famous/core/Surface');
 var template        = require('./Menu.jade');
-var TapHandler      = require('../helpers/TapHandler');
+var TapHandler      = require('../../helpers/TapHandler');
 
-function PageView(options) {
+function MenuBtnView(options) {
     View.apply(this, arguments);
 
     var mod = new Modifier({
-      transform: Transform.translate(0,0,1)
+      transform: Transform.translate(0,0,1),
+      size: [undefined,30]
     });
 
     var surface = new Surface({
-      content: template(options.content),
-      classes: ['menu']
+      content: options.author,
+      classes: ['menu-btn']
     });
     surface.on('tap',TapHandler);
 
     this.add(mod).add(surface);
 }
 
-PageView.prototype = Object.create(View.prototype);
-PageView.prototype.constructor = PageView;
+MenuBtnView.prototype = Object.create(View.prototype);
+MenuBtnView.prototype.constructor = MenuBtnView;
 
-PageView.DEFAULT_OPTIONS = {};
+MenuBtnView.DEFAULT_OPTIONS = {};
 
-module.exports = PageView;
+module.exports = MenuBtnView;